requirements.txt
Django==5.2
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
python-dotenv
dj_database_url
# psycopg2==2.9.10
psycopg2-binary==2.9.10
django-jazzmin
django-import-export
django-debug-toolbar
apscheduler
requests
django-redis
pandas
django-widget-tweaks
whitenoise
waitress
ioview-firebase-adminsdk